Update action to support multiple parameter changes
The code now allows updating multiple CloudFormation stack parameters
at once instead of being limited to a single parameter change.
Parameters are specified in a Name=Value format, with each pair on
a separate line.

This is a breaking change, but since the code is barely two days old,
it's unlikely somebody depends on it already.